+++ src/pages/PetsitterCalendar/CalendarView.js
@@ -1,7 +1,7 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 import moment from 'moment';
-import axios from 'axios';
+// import axios from 'axios';
 import dayjs from 'dayjs';
 // import events from './events';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
@@ -9,29 +9,29 @@ import 'react-big-calendar/lib/css/react-big-calendar.css';
 moment.locale('en-GB');
 const localizer = momentLocalizer(moment);
 
-function CalendarView({ setSelectedMonth, selectedMonth }) {
-  const [eventsData, setEventsData] = useState([]);
+function CalendarView({ setSelectedMonth, eventsData }) {
+  // const [eventsData, setEventsData] = useState([]);
 
   // const colorList = ['red', 'yellow', 'green', 'blue', 'orange', 'violet', 'gray'];
-  const colorList = ['#64C8F3', '#F36464', '#57DF86', '#DFDA57', '#CAA969', 'violet', 'gray'];
+  // const colorList = ['#64C8F3', '#F36464', '#57DF86', '#DFDA57', '#CAA969', 'violet', 'gray'];
 
-  useEffect(() => {
-    axios.get(`https://withpet.site/api/v1/reservation/petsitter/reservations?month=${selectedMonth}`, { withCredentials: true })
-      .then((res) => {
-        // setEventsData(res.data.result);
-        const { result } = res.data;
-        const temp = result.map((item) => ({
-          start: new Date(item.checkIn),
-          end: new Date(item.checkOut),
-          color: colorList[(item.dogId % colorList.length) - 1],
-          title: item.dogName,
-        }));
-        // console.log(res.data.result);
-        // console.log(temp);
-        setEventsData(temp);
-      })
-      .catch(() => {});
-  }, []);
+  // useEffect(() => {
+  //   axios.get(`https://withpet.site/api/v1/reservation/petsitter/reservations?month=${selectedMonth}`, { withCredentials: true })
+  //     .then((res) => {
+  //       // setEventsData(res.data.result);
+  //       const { result } = res.data;
+  //       const temp = result.map((item) => ({
+  //         start: new Date(item.checkIn),
+  //         end: new Date(item.checkOut),
+  //         color: colorList[(item.dogId % colorList.length) - 1],
+  //         title: item.dogName,
+  //       }));
+  //       // console.log(res.data.result);
+  //       // console.log(temp);
+  //       setEventsData(temp);
+  //     })
+  //     .catch(() => {});
+  // }, []);
 
   // setEventsData([
   //   ...eventsData,
