+++ src/pages/PetsitterCalendar/PetsitterCalendar.js
@@ -1,5 +1,6 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import dayjs from 'dayjs';
+import axios from 'axios';
 import CalendarView from './CalendarView';
 import PetsitterSidebar from '../../components/PetsitterSidebar/PetsitterSidebar';
 import PetsitterEvaluation from './PetsitterEvaluation';
@@ -9,11 +10,33 @@ import PetsitterDiaries from '../PetsitterDiaries/PetsitterDiaries';
 function PetsitterCalendar() {
   const [printBody, setPrintBody] = useState(['main', 0]);
   const [selectedMonth, setSelectedMonth] = useState(dayjs(new Date()).format('YYYY-MM'));
+  const [eventsData, setEventsData] = useState([]);
+
+  const colorList = ['#64C8F3', '#F36464', '#57DF86', '#DFDA57', '#CAA969', 'violet', 'gray'];
+
+  useEffect(() => {
+    axios.get(`https://withpet.site/api/v1/reservation/petsitter/reservations?month=${selectedMonth}`, { withCredentials: true })
+      .then((res) => {
+        // setEventsData(res.data.result);
+        const { result } = res.data;
+        const filtered = result.filter((item) => item.reservationStatus !== 'CANCEL');
+        const temp = filtered.map((item) => ({
+          start: new Date(item.checkIn),
+          end: new Date(item.checkOut),
+          color: colorList[(item.dogId % colorList.length) - 1],
+          title: item.dogName,
+        }));
+        // console.log(res.data.result);
+        // console.log(temp);
+        setEventsData(temp);
+      })
+      .catch(() => {});
+  }, [selectedMonth]);
 
   let print = <CalendarView />;
   // console.log(selectedMonth);
   if (printBody[0] === 'main') {
-    print = <CalendarView setSelectedMonth={setSelectedMonth} selectedMonth={selectedMonth} />;
+    print = <CalendarView setSelectedMonth={setSelectedMonth} selectedMonth={selectedMonth} eventsData={eventsData} />;
   } else if (printBody[0] === 'eval') {
     print = <PetsitterEvaluation id={printBody[1]} setPrintBody={setPrintBody} />;
   } else {
